+++ Packages/Tracking/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	- Interaction Cube, 3D Button, 3D UI panel, Anchor, Anchorable Object, Attachment Hand Menu, Interaction Manager
 - (Physics Hands) Added function to check if a grasped object has been pinched
 - Option to show Upper Arm for Capsule Hands. (Works best in XR)
+- (Preview) XRI implementation now supports more Input Actions similar to that of OpenXRs Interaction Profiles
 
 ### Changed
 - (Anchorable Behaviour) Ability to change the speed at which an object is attracted to the hand